Add GitHub Action workflow to upload releases
- Uploads distribution files as GA artifacts
- Uploads to GitHub Releases
- Uploads to TestPyPI (for now)

As of this revision, `nox -s release` uploads to PyPI and pushes the tag
to the GitHub repository. This change generates the release artifacts in
a GitHub Action workflow and uploads them to various places that
`nox -s release` does not upload to.

This allows this workflow to co-exist with the current functional
release pipeline, thus allowing for testing and tweaks before making a
"real" release using this mechanism.