Target [source] links more accurately in SASS
Because `.viewcode-link` is already marked with `float: right`
independently, don't assign that style redundantly when styling
`[source]` links attached to function and class defintions.

Use `.viewcode-link` instead of `a.reference.internal` to ensure that
the styling for `[source]` links is not accidentally applied to return
type annotations which are also internal references.