Add a test case to kitchen-sink for autodoc
In conf.py, modify sys.path to add `kitchen_sink/demo_py/`
This is the approach taken in the sphinx_rtd_theme in order to add an
autodoc'ed module [1].

This module contains a class with one method and a function. This is
more than enough to ensure that autodoc with type annotated methods
and functions is exercised in furo.

[1] https://github.com/readthedocs/sphinx_rtd_theme/blob/671703c3fbb5629fb5004afc941b59310976c28b/docs/conf.py#L10