Fallback to project name branding/title for singlehtml
The title element and sidebar branding name both use the `docstitle`
configuration [1][2] for its content; however, when using a `singlehtml`
builder, this option is not available [1][3]. In the case where this
option is not available, this will lead to the sidebar name rendering a
value "None" (unless `sidebar_hide_name` is explicitly configured to
`True`) and some title values as "None". This commit adjusts the case
where `docstitle` is not configured, by falling back onto the configured
`project` value (for the sidebar) or the `title` value (for the title
element).

[1]: https://www.sphinx-doc.org/en/master/templating.html?highlight=docstitle#docstitle
[2]: https://github.com/sphinx-doc/sphinx/blob/514fca7a407f03fae4c788178555a74256936655/sphinx/builders/html/__init__.py#L509
[3]: https://github.com/sphinx-doc/sphinx/blob/514fca7a407f03fae4c788178555a74256936655/sphinx/builders/singlehtml.py#L139

Signed-off-by: James Knight <james.d.knight@live.com>