Rework the asset generation setup
Switching to sphinx-theme-builder simplifies the Python packaging
process, by integrating the NodeJS build step within the Python wheel
generation step. This creates much simpler semantics around installation
by avoiding the need for a separate NodeJS build step outside of the
Python installation process.

Grunt enables a simpler format for describing builds, making it much
easier to reason about and set up. This is especially relevant since
the builds are now executed with NodeJS 16 and most of the Gulp plugins
in the previous setup were not ready for Node 16.