Force highlight (pre) line height style
When toggling between light and dark mode, Pygments will configure a
line-height value that is inconsistent with the existing light mode.
The code style configures a `line-height` of `1.5`, but when Pygments
styling is changed, the line-height is adjusted to `125%`:

    @media (prefers-color-scheme: dark)
    body:not([data-theme="light"]) .highlight pre {
        line-height: 125%;
    }

To prevent this, providing an important hint on the line-height entry to
ensure a consistent height between the two modes.

Signed-off-by: James Knight <james.d.knight@live.com>