Improve support for print media
The following improves the ability to print documentation themed with
Furo. This includes hiding various elements (such as the mobile bar)
that are not useful in printed media, as well as deals with color
scheming issues for browsers who do not default to light scheme in a
print media state.

Signed-off-by: James Knight <james.d.knight@live.com>