ultra-basic PoC
This "works" with sphinx-multiversion.

It's very ugly, but the versions *will* show up in the sidebar if we're
not being built on RTD. From Salt's perspective we probably want to just
keep with the recent approach of master/main/latest, and then adding the
previous two versions. Or maybe even just limiting build to the
supported versions?