Add block to allow inserting content before `main`
When extending Furo using a template with `extends`, there's currently
not a way to add content within the `page` but before or after `main`.
An example would be a top navigation bar that sits above the rest of the
content but not over the left-side navigation bar.

This change adds a block that allows this through a template like this:
```
{% extends "furo/page.html" %}

{% block supermain %}
  <div class="my-custom-content">
    <h1>Hello World</h1>
  </div>
  {{ super() }}
  <div class="my-custom-content-on-the-bottom">
    <h1>Goodbye World</h1>
  </div>
{% endblock supermain %}
```
If you don't use templates to extend Furo, or if your templates don't
define the `supermain` block, this change has no effect.